# Conflicts:
#   src/pages/admin/ProductManagement.jsx
#   src/services/ProductAPI.js

+++ src/pages/admin/UserManagement.jsx
@@ -1,28 +1,16 @@
 import React, { useState, useEffect } from 'react';
-import { getUserList } from '../../services/AdminAPI';
+import { getUserList } from '@/services/AdminAPI';
 import { Box, Heading, Table, Button, HStack, Stack, Input, createListCollection } from '@chakra-ui/react';
 import { Toaster, toaster } from '@/components/ui/toaster';
 import { Checkbox } from '@/components/ui/checkbox'
-import {
-    SelectContent,
-    SelectItem,
-    SelectRoot,
-    SelectTrigger,
-    SelectValueText,
-} from '@/components/ui/select'
-import {
-    PaginationItems,
-    PaginationNextTrigger,
-    PaginationPrevTrigger,
-    PaginationRoot,
-} from '@/components/ui/pagination'
+import { SelectContent, SelectItem, SelectRoot, SelectTrigger, SelectValueText } from '@/components/ui/select'
+import { PaginationItems, PaginationNextTrigger, PaginationPrevTrigger, PaginationRoot } from '@/components/ui/pagination'
 import { useSearchParams } from 'react-router-dom';
-import { ToggleTip } from "@/components/ui/toggle-tip"
-import { LuInfo } from "react-icons/lu"
-import { Switch } from "@/components/ui/switch"
-import { HiCheck, HiX } from "react-icons/hi"
-import { updateLoginLock, updateAuth, deleteUser, deleteUsers } from '../../services/AdminAPI';
-import { color } from 'framer-motion';
+import { ToggleTip } from '@/components/ui/toggle-tip'
+import { LuInfo } from 'react-icons/lu'
+import { Switch } from '@/components/ui/switch'
+import { HiCheck, HiX } from 'react-icons/hi'
+import { updateLoginLock, updateAuth, deleteUser, deleteUsers } from '@/services/AdminAPI';
 
 function UserManagement() {
     const [users, setUsers] = useState([]);
@@ -83,12 +71,17 @@ function UserManagement() {
     };
 
     const handleDeleteUser = (email) => {
-        deleteUser(email)
-            .then(() => {
-                loadUsers();
-            })
-            .catch((error) => {
-            });
+        const confirmed = window.confirm('삭제하시겠습니까?');
+        if (confirmed) {
+            deleteUser(email)
+                .then(() => {
+                    loadUsers();
+                })
+                .catch((error) => {
+                });
+        } else {
+            alert('취소되었습니다.');
+        }
     };
 
     const handleDeleteUsers = () => {
@@ -96,13 +89,17 @@ function UserManagement() {
             alert('삭제할 사용자를 선택하세요.');
             return;
         }
-        
-        deleteUsers(selection)
-            .then(() => {
-                loadUsers();
-            })
-            .catch((error) => {
-            });
+        const confirmed = window.confirm('삭제하시겠습니까?');
+        if (confirmed) {
+            deleteUsers(selection)
+                .then(() => {
+                    loadUsers();
+                })
+                .catch((error) => {
+                });
+        } else {
+            alert('취소되었습니다.');
+        }
     };
 
     const formatDate = (dateString) => {
@@ -113,11 +110,15 @@ function UserManagement() {
     return (
         <Box>
             <Toaster />
-            <Box display='flex' justifyContent='space-between' alignItems='center' mb='3'>
+            <Box display='flex' justifyContent='space-between'          
+                alignItems='center' mb='3'
+            >
                 <HStack gap='0' alignItems='center'>
                     <Heading as='h1' size='xl' mr='5'>회원 관리</Heading>
-                    <SelectRoot size='sm' width='110px' collection={types} onChange={(e) => setType(e.target.value)}
-                    style={{ height: '36px' }}
+                    <SelectRoot size='sm' width='110px'
+                        collection={types} 
+                        onChange={(e) => setType(e.target.value)}
+                        style={{ height: '36px' }}
                     >
                         <SelectTrigger>
                             <SelectValueText placeholder='검색 유형' />
@@ -139,12 +140,18 @@ function UserManagement() {
                     />
                 </HStack>
                 <HStack>
-                    <ToggleTip content="선택한 회원을 탈퇴 시킬 수 있습니다.">
-                                <Button size="xs" variant="ghost">
-                                    <LuInfo />
-                                </Button>
+                    <ToggleTip
+                        content='선택한 회원을 탈퇴 시킬 수 있습니다.'
+                    >
+                        <Button size='xs' variant='ghost'>
+                            <LuInfo />
+                        </Button>
                     </ToggleTip>
-                    <Button colorScheme='red' onClick={() => handleDeleteUsers()}>탈퇴</Button>
+                    <Button
+                        onClick={() => handleDeleteUsers()}
+                    >
+                        탈퇴
+                    </Button>
                 </HStack>
             </Box>
             <Box borderBottom={{ base: '1px solid black', _dark: '1px solid white' }} mb='3' />
@@ -169,17 +176,22 @@ function UserManagement() {
                             <Table.ColumnHeader fontSize='md'>이름</Table.ColumnHeader>
                             <Table.ColumnHeader fontSize='md'>가입일</Table.ColumnHeader>
                             <Table.ColumnHeader fontSize='md'>로그인
-                                <ToggleTip content="
-                                로그인 잠금 상태를 변경 할 수 있습니다.">
-                                    <Button size="xs" variant="ghost">
+                                <ToggleTip
+                                    content='로그인 잠금 상태를 변경 할 수 있습니다.'
+                                >
+                                    <Button
+                                        size='xs'
+                                        variant='ghost'
+                                    >
                                         <LuInfo />
                                     </Button>
                                 </ToggleTip>
                             </Table.ColumnHeader>
                             <Table.ColumnHeader fontSize='md'>관리자
-                                <ToggleTip content="회원 권한을 변경 할 수 있습니다."
+                                <ToggleTip
+                                    content='회원 권한을 변경 할 수 있습니다.'
                                 >
-                                    <Button size="xs" variant="ghost">
+                                    <Button size='xs' variant='ghost'>
                                         <LuInfo />
                                     </Button>
                                 </ToggleTip>
@@ -189,7 +201,9 @@ function UserManagement() {
                     </Table.Header>
                     <Table.Body>
                         {(users.length > 0 ? users : []).map((user) =>
-                            <Table.Row key={user.email} data-selected={selection.includes(user.email) ? '' : undefined}>
+                            <Table.Row
+                                key={user.email}
+                                data-selected={selection.includes(user.email) ? '' : undefined}>
                                 <Table.Cell>
                                     <Checkbox
                                         top='1'
@@ -197,9 +211,8 @@ function UserManagement() {
                                         checked={selection.includes(user.email)}
                                         onCheckedChange={(changes) => {
                                             setSelection((prev) =>
-                                                changes.checked
-                                                    ? [...prev, user.email]
-                                                    : selection.filter((email) => email !== user.email),
+                                                changes.checked ? [...prev, user.email]
+                                                : selection.filter((email) => email !== user.email),
                                             )
                                         }}
                                     />
@@ -210,33 +223,46 @@ function UserManagement() {
                                 <Table.Cell>{formatDate(user.createdAt)}</Table.Cell>
                                 <Table.Cell>
                                     <Switch
-                                        size="lg"
-                                        variant="raised"
+                                        size='lg'
+                                        variant='raised'
                                         checked={user.loginLock==true}
                                         onChange={() => handleUpdateLoginLock(user.email)}
-                                        thumbLabel={{ on: <HiCheck color="white" />,
-                                                    off: <HiX /> }}>
+                                        thumbLabel={{ 
+                                            on: <HiCheck color='white' />,
+                                            off: <HiX />
+                                        }}>
                                     </Switch>
                                 </Table.Cell>
                                 <Table.Cell>
                                     <Switch
-                                        size="lg"
-                                        variant="raised"
+                                        size='lg'
+                                        variant='raised'
                                         checked={user.userAuth == 'ROLE_ADMIN'}
                                         onChange={() => handleUpdateAuth(user.email)}
-                                        thumbLabel={{ on: <HiCheck color="white" />,
-                                                    off: <HiX /> }}>
+                                        thumbLabel={{ 
+                                            on: <HiCheck color='white' />,
+                                            off: <HiX />
+                                        }}>
                                     </Switch>
                                 </Table.Cell>
                                 <Table.Cell>
-                                    <Button onClick={() => handleDeleteUser(user.email)}>탈퇴</Button>
+                                    <Button
+                                        onClick={() => handleDeleteUser(user.email)}
+                                    >
+                                        탈퇴
+                                    </Button>
                                 </Table.Cell>
                             </Table.Row>
                         )}
                     </Table.Body>
                 </Table.Root>
             </Box>
-            <Stack gap='4' alignItems='center' mt='3' mb='3'>
+            <Stack
+                gap='4'
+                alignItems='center'
+                mt='3'
+                mb='3'
+            >
                     <PaginationRoot
                         page={page}
                         count={totalPages}
